fixed types for json and log error message in LogWatcherService

+++ src/Models/Pool/PoolStatistics.cs
@@ -3,9 +3,9 @@
 public class PoolStatistics
 {
     public double Diff { get; set; }
-    public int Accepted { get; set; }
-    public int Rejected { get; set; }
-    public double BestShare { get; set; }
+    public long Accepted { get; set; }
+    public long Rejected { get; set; }
+    public decimal BestShare { get; set; }
     public double SPS1m { get; set; }
     public double SPS5m { get; set; }
     public double SPS15m { get; set; }
